First Iteration with Poynting Magnitudes
Sums the Poynting vector magnitudes for each point in the specified
recti-formed space. The output is the final sum.
Accounting for all three dimensions allows for the greatest utility. In
order to accomodate a 1- or 2-dimensional simulation, simply specify D =
1 or D = 2 and set to 0 the input field arrays that are not used in the
simulation.
Note that currently, the function computes the vector magnitudes for the
Simulation-points, which are not the same as the intended
Evaluation-points, which would be averages of the Simulation-points.