First Iteration Poynting Magnitudes
Sums the Poynting vector magnitudes of each point in an input-specified
recti-formed space. The output is the final sum.

Accounting for three dimensions allows for the greatest utility. In
order to accomodate 1- or 2-dimensional simulations, set D = 1 or D = 2
and set to 0 the unsimulated input field-arguments.

Note that currently, the function computes the vector magnitudes of the
Simulation-points, not the Evaluation-points, which would be averages of
the Simulation-points.