Conversion to Static Array Storage
Initializing the field and PML arrays in static memory (without allocate
calls) increases the amount of memory available for those arrays
(regular memory instead of dynamic memory).

In order to accomodate this change, each array must possess the storage
potential equal to that required for the largest resolution. This does
not violate any FDTD rules, as the extra 0's are simply never updated
and will not affect the other array values.

Note that my absoft compiler yield syntax errors for the PML declaration
section of "Field and PML Arrays."