Added Relevant Codes
These "Examples" are pieces of code that either (are or include)
features that are important to Convergence_Vacuum. Grid_Return has been
left as an entire code, but the only parts intented for use in
Convergence_Vacuum are the Grid-Return lines. I left the rest of the
code intact because the position of each Grid-Return block relative to
the main structure is also important.