Horrible Mistake Spotted!!!
Nx and Ny were swapped! the physical values corresponded to the wrong
computer values!

All previous runs of Convergence were INCORRECT due to this vital
mistake.

On the bright side, this probably indicates that a buffer is not
required; however, I'll keep it there anyways, since it is the most
general version of a finder.